+++ src/main/kotlin/smalltalk/backend/infrastructure/repository/room/RoomRepository.kt
@@ -1,15 +1,15 @@
 package smalltalk.backend.infrastructure.repository.room
 
-import org.springframework.stereotype.Repository
 import smalltalk.backend.domain.room.Room
+import java.math.BigInteger
 
 interface RoomRepository {
-    fun save(roomName: String): Long?
-    fun findById(roomId: Long): Room?
+    fun save(roomName: String): BigInteger?
+    fun findById(roomId: BigInteger): Room?
     fun findAll(): List<Room>
     fun addMember(room: Room): Room
-    fun deleteMember(roomId: Long, memberId: Long)
+    fun deleteMember(roomId: BigInteger, memberId: BigInteger)
     fun updateRoom(room: Room)
-    fun deleteById(roomId: Long): Boolean
+    fun deleteById(roomId: BigInteger): Boolean
     fun deleteAll(): Long
 }
